refactor
# Conflicts:
#	react/src/mapper/map-generator.tsx
#	react/src/urban-stats-script/constants/constants.ts
#	react/src/urban-stats-script/derive-human-readable-name.ts
#	react/src/urban-stats-script/human-readable-template.ts
#	react/unit/urban-stats-script-derive-human-readable-name.test.ts
