+++ react/src/components/article-panel.tsx
@@ -10,8 +10,8 @@ import { groupYearKeys, StatGroupSettings } from '../page_template/statistic-set
 import { statParents } from '../page_template/statistic-tree'
 import { PageTemplate } from '../page_template/template'
 import { Universe, universeContext, useUniverse } from '../universe'
-import { HumanReadableName } from '../urban-stats-script/types-values'
 import { assert } from '../utils/defensive'
+import { HumanReadableName } from '../utils/human-readable-name'
 import { sanitize } from '../utils/paths'
 import { Article, IRelatedButtons } from '../utils/protos'
 import { useComparisonHeadStyle, useHeaderTextClass, useMobileLayout, useSubHeaderTextClass } from '../utils/responsive'
