+++ react/src/mapper/settings/Selector.tsx
@@ -9,12 +9,12 @@ import { UrbanStatsASTExpression } from '../../urban-stats-script/ast'
 import { hsvToColor, rgbToColor } from '../../urban-stats-script/constants/color'
 import { Color, doRender, hexToColor, hsvColorExpression, rgbColorExpression } from '../../urban-stats-script/constants/color-utils'
 import { RampT } from '../../urban-stats-script/constants/ramp'
-import { reifyReact, reifyString } from '../../urban-stats-script/derive-human-readable-name'
 import { EditorError } from '../../urban-stats-script/editor-utils'
 import { emptyLocation } from '../../urban-stats-script/lexer'
 import { parseNoErrorAsCustomNode, parseNoErrorAsExpression } from '../../urban-stats-script/parser'
-import { Documentation, HumanReadableName, TypeEnvironment, USSType } from '../../urban-stats-script/types-values'
+import { Documentation, TypeEnvironment, USSType } from '../../urban-stats-script/types-values'
 import { TestUtils } from '../../utils/TestUtils'
+import { HumanReadableName, reifyReact, reifyString } from '../../utils/human-readable-name'
 import { useTextAreaSizeSync } from '../../utils/text-area-size-sync'
 
 import * as l from './../../urban-stats-script/literal-parser'
