+++ react/src/stat/StatisticPanelTable.tsx
@@ -9,9 +9,9 @@ import { Navigator } from '../navigation/Navigator'
 import { useColors } from '../page_template/colors'
 import { useUniverse } from '../universe'
 import { orderNonNan } from '../urban-stats-script/constants/table'
-import { reifyString } from '../urban-stats-script/derive-human-readable-name'
 import { TypeEnvironment } from '../urban-stats-script/types-values'
 import { assert } from '../utils/defensive'
+import { reifyString } from '../utils/human-readable-name'
 import { sanitize } from '../utils/paths'
 
 import { makeColumnDeleteHandler } from './makeColumnDeleteHandler'
