+++ react/src/urban-stats-script/ast.ts
@@ -1,5 +1,7 @@
 import assert from 'assert'
 
+import { HumanReadableElement } from '../utils/human-readable-name'
+
 import { AutoUXNodeMetadata } from './autoux-node-metadata'
 import { LocInfo } from './location'
 import { BinaryOperatorSymbol, UnaryOperatorSymbol } from './operators'
