+++ react/src/urban-stats-script/types-values.ts
@@ -1,6 +1,7 @@
 import type * as ArgEditButtons from '../mapper/settings/ArgEditButtons'
 import { Basemap } from '../mapper/settings/utils'
 import { assert } from '../utils/defensive'
+import { HumanReadableName } from '../utils/human-readable-name'
 import { UnitType } from '../utils/unit'
 
 import { UrbanStatsASTExpression } from './ast'
@@ -12,7 +13,6 @@ import { Scale } from './constants/scale'
 import { Table, TableColumn } from './constants/table'
 import { TextBox } from './constants/text-box'
 import { Context } from './context'
-import { HumanReadableElement } from './derive-human-readable-name'
 import { ConstantCategory } from './documentation-category'
 import { noLocation } from './location'
 import { unparse } from './parser'
@@ -141,8 +141,6 @@ export type USSRawValue = (
 export type DocumentationTable = 'mapper-data-variables' | 'predefined-colors' | 'unit-types' | 'predefined-ramps' | 'predefined-insets' | 'logarithm-functions' | 'trigonometric-functions'
 
 export type SelectorRendering = { kind: 'subtitleLongDescription' } | { kind: 'gradientBackground', ramp: RampT }
-
-export type HumanReadableName = string | HumanReadableElement[]
 
 export interface Documentation {
     humanReadableName: string
