+++ react/src/utils/human-readable-name.tsx
@@ -2,6 +2,8 @@ import React, { ReactNode } from 'react'
 
 export type HumanReadableElement = { type: 'atom', value: string } | { type: 'where' | 'superscript' | 'subscript' | 'parens', value: HumanReadableElement[] }
 
+export type HumanReadableName = string | HumanReadableElement[]
+
 export function reifyReact(elements: HumanReadableElement[] | string): ReactNode {
     if (typeof elements === 'string') return elements
     return elements.map((element, index) => {
